LPS-73481 Move upgrade processes implemented after the release of GA3 to a new upgrade version
Conflicts:
	portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_4.java
