+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_4/UpgradeContact.java
@@ -12,20 +12,20 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_0_3;
+package com.liferay.portal.upgrade.v7_0_4;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.upgrade.v7_0_3.util.UserTable;
+import com.liferay.portal.upgrade.v7_0_4.util.ContactTable;
 
 /**
  * @author Ugurcan Cetin
  */
-public class UpgradeUser extends UpgradeProcess {
+public class UpgradeContact extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		alter(
-			UserTable.class,
+			ContactTable.class,
 			new AlterColumnType("emailAddress", "VARCHAR(254) null"));
 	}
